Update Peer-graded Assignment: Peer Assignment: Build a Circuit
Build a circuit that contains two pushbuttons, an LED, and any other basic components you think you need. The LED should turn on when either the first button or the second button is pressed. Only use your Arduino for power and ground.

Create a video (or a screencast) of your circuit in action.

If you have real hardware then demonstrate the real circuit. If you do not have real hardware then use the simulator.

Regardless of which method you use, provide a link to the location of your video or screencast on the Internet. Do not upload your video here.

Show the behavior of the system under 3 conditions:

1.no buttons pressed

2.button 1 pressed alone

3.button 2 pressed alone